Use epsilon comparison in assigning points
If the point cloud is very small, the hull got into an inconsistent
state. Reassigning with epsilon comparison fixes that. Basically, if a
point is less than an epsilon above the triangle being assigned, it's
either behind it or considered coplanar with the triangle. Either way,
it's considered inside will not be considered when growing the hull.